// src/game/logic/types/enums/Traits.enum.ts
export enum TRAITS_ENUM {
    SHARPSHOOTER = "Sharpshooter",
    TOUGH = "Tough",
    FAST_LEARNER = "Fast Learner",
    HANDY = "Handy",
    GREEN_THUMB = "Green Thumb",
    HEALER = "Healer",
    CHARISMATIC = "Charismatic",
    ANIMAL_WHISPERER = "Animal Whisperer",
    MINER = "Miner",
    SCIENTIST = "Scientist",
    ATHLETE = "Athlete",
    ARTISAN = "Artisan",
    SNEAKY = "Sneaky",
    LUCKY = "Lucky",
    TECHNICIAN = "Technician",
    DIPLOMAT = "Diplomat",
    MUSICIAN = "Musician",
    NURTURER = "Nurturer",
    MYSTERIOUS = "Mysterious",
    JOKER = "Joker",
}